% Please edit documentation in R/demo1-GEN.R, R/demo1_easier-GEN.R
\name{add_one}
\alias{add_one}
\title{A cool function}
\usage{
add_one(x)

add_one(x)
}
\arguments{
\item{x}{a numeric vector}

\item{x}{a numeric vector}
}
\description{
Well, not really cool. Just add 1 to x.

Well, not really cool. Just add 1 to x.
}
\examples{
add_one(1)
add_one(1:10)
add_one(1)
add_one(1:10)
}
